automotive-image-runner: require netcat if --publish-dir is specified
If --publish-dir is specified on the command line, and netcat is not
found in the path, then a warning message is only shown for a fraction
of a second on the screen. This will make it so that updates won't work.
Let's make this a fatal error since the intent of the user is to make
updates available to the VM.

Signed-off-by: Brian Masney <bmasney@redhat.com>