Work around rpm-ostree selinux bug
rpm-ostree needs to add some stuff to
/etc/selinux/$policy/contexts/files/file_contexts.subs_dist, but
before https://github.com/coreos/rpm-ostree/pull/5382 this incorrectly
hardcoded the targeted policy.

This is now fixed, but the fix is not yet in cs9. For now we work
around this, at least for the critical /usr/etc, by doing this
ourselves.

This seems to not be critical for normal use of rhivos, but it completely
breaks use with bootc if we don't do this.