Drop "osname" option everywhere, us "default" to match bootc
With bootc-image-builder, the stateroot name defaults to "default", and
it isn't possible for the image to override this. We currently use
a distro name for this, but that makes no real sense. All deploys in a
car should use the same stateroot, and it doesn't have to have the distro
name in it.

Lets sync with what bootc and just always default to "default".