Change default boot mechanism to ukiboot
Lets try this by default now, as it has been added to autosd and
rhivos. We still default to grub2 when building bootc images, as that
is not yet supported.

It is expected that this may not work in all cases, like for example
specific hardware targets, so for now we may need further changes to
explicitly use grub for those.

All targets that use DOS partitioning (instead of GPT) are set to use
grub2, as ukiboot requires GPT. Maybe these can be hybrid instead later.