Enable ostree_booted when installing rpms to fix nss-altfiles
In cs10, the nsswitch.conf file is generated by the authselect
post-transaction scriptlet. This needs to know if ostree is used or
not because then it enables nss-altfiles (which we need). To
handle this we ensure ostree_booted is set to true in the rpm
install stage in image mode.