automotive-image-runner: Default to virtio-gpu display
This is a much more modern and nicer display mechanism compared to the
traditional default we've used (vga on x86, nothing anywhere else).

Some advantages:
 * This works on arm as well as x86
 * When using virtio_gpu, window resize trigger monitor resolution
   changes event, which some setups can use to resize the console.

If vga support is really needed (say for a kernel that doesn't
have virtio-gpu support, use the "-vga" commandline option.

Also, this is currently the only way to get a console on arm with cs10,
because the arm serial console driver is not built in atm.