feat(crossref): Use crossref for mpp imported files
Use git-crossref [1] to synchronize utility mpp files
from upstream repositories (i.e., ostream).
The configuration is defined in `.gitcrossref`
in the repository root.

Using `git-crossref` has the advantage of tracking
which hash/version/branch the files are synced with.
And also has a handy check subcommand to ensure that
the files are not locally modified. This check is
added as a gitlab job and would fail if anyone
manually modifies the files.

To install locally:

```bash
pip install git-crossref
`````

To update the synchronized files from upstream repositories:

```bash
git-crossref sync
`````

This command will fetch the latest versions of files specified in the
configuration and apply any defined transformations.

To modify which files are synchronized or update upstream versions:

1. Edit the `.gitcrossref` file
2. Update the `version` field to point to a specific commit,
   version or branch
3. Add or remove files in the `files` section as needed
4. Run `git-crossref sync` to apply the changes

[1] - https://github.com/aesteve-rh/git-crossref

Signed-off-by: Albert Esteve <aesteve@redhat.com>