Raise the inability to obtain host from AWS
Currently the tmt integratiomn test script continues to run the tests
even after there was an error around inability to obtain a host from
AWS, which made this error pretty hidden in the logs. This fix exits the
script with an error right after it cannot get a host from AWS pool.

Signed-off-by: Martin Perina <mperina@redhat.com>