Added ShutdownHandler to automotive-image-runner
Executing a-i-r with the --publish-dir option will setup an http server
used to exchange data, i.e. ostree updates, between the host and the VM.
Stopping a-i-r with ctrl + c (sigint) results in an unhandled exception
in the forked process, stopping the http server. In order to provide a
smooth shutdown, a handler reacting on sigint and sigterm has been added.

Signed-off-by: Michael Engel <mengel@redhat.com>