Add a-i-b commands bootc-to-disk-image and build-bootc-builder
The `build-bootc-builder` command is subset of `build` which builds a
container that can be used by bootc-image-builder. It keeps a subset
of the options that `build` has, and passes fixed options to the rest.

The `bootc-to-disk-image` command takes a bootc container image from
the local container store and runs it through bootc-image-builder to
produce a disk image. It takes as additional options a build container
(you can create one with the `build-bootc-builder` command, or use a
prebuilt one), and a bootc-image-builder container (defaults to
`quay.io/centos-bootc/bootc-image-builder:latest`).

For example, these commands will build and run an image:

```
$ automotive-image-builder build --export bootc examples/simple.aib.yml localhost/simple:latest
$ automotive-image-builder build-bootc-builder localhost/auto-bootc-build-autosd9-sig:latest
$ automotive-image-builder bootc-to-disk-image localhost/simple:latest disk.img
$ automotive-image-runner disk.img
```