Fix swapped root partition UUIDs for x86_64/aarch64
Dana Elfassy reports that images built on x86_64 machines contain rootfs
partitions with a partition type of "Linux root (ARM-64)".

This happens due the rootfs partition type UUIDs in computed-vars.ipp.yml
being swapped. The UUID for x86_64 (4f68...) was assigned to the aarch64
key, and the UUID for aarch64 (b921...) was assigned to the x86_64 key.

Swap the UUIDs to align them with their correct architectures as defined
by the Discoverable Partitions Specification.