Add lazy variables for kernel evr/evra and kernel modules this
During the parsing of include files we can't access the rpms variable
which contains the result of the depsolve, because it is computed after
completing the include (but before starting to format the resulting
pipelines/stages.

This adds a new _Lazy class that allows the new variables to be defined
in the standard way, but then resolved later when used in the stages.