Use org.osbuild.write-device instead of automotive specific one
Now that we have an upstream write-device, use it instead of the
osbuild-auto one.

Note that the osbuild-auto one had some weird magic to look at the
ostree deploy directory if the specified file didn't exist. This
is not in the upstream version, so we instead use a pre-deploy
tree to pick up the file in the standard location.

Also, this merges the aboot and ukiboot calls, as they are essentially
the same.