Fix the selinux label alias for /usr/etc/systemd/system
This is fixed upstream in https://github.com/coreos/rpm-ostree/pull/5485
but lets carry the workaround here until that is available.

This is important, because without this fix then relabeling in bootc
after having moved /etc to /usr/etc will produce different results
for some files, which makes the ostree commit that bootc produces for
the container slightly different, which makes secure-boot style content
validation impossible.