ukiboot: Replace the shim EFI binary in both paths
The shim binary is installed in two places: 1) the distro EFI directory
(\EFI\centos) and 2) the default boot path (\EFI\BOOT). The former is
used as the path in the boot entry for the distribution and the latter
is used as a default path if there are no boot entries in the EFI NVRAM.

We need to replace both paths to ensure that the ukiboot EFI binary will
be always executed regardless of the EFI variables present in the system.