Switch to using root=UUID=fs_uid to find root
Using the fs id to find the root filesystem is the most general
approach that works in all cases. For example, MBR doesn't have
partition uuids or labels, so those don't work. And when using LUKS or
other kind of more complex setups, the partition id is also not
enough.

For cases where we need to use a custom approach for boot performance
reasons. Like say if there are many partitions which make enumeration
slow, then the target can set use_autoinit_root to true which default
to by PARTUUID, but that can be overridden using autoinit_root.

When autoinit_root is used we still *also* pass root=UUID=... to
ensure that userspace has a consistent standard kernel cmdline.