Add --build-dir pointing to a non-overlayfs directory
Otherwise, by default build dir is in /var/tmp, and when running in CI
this is a standard container location, which is mounted on an overlayfs.
When then osbuild tries to create a new container in some subdirectory
of this it tries to mount a new overlayfs with the lower directory being
on this overlayfs, and overlayfs doesn't support such recursive use.