feat(wayland): Add flag to define compositor image
Add qm_wayland_enable flag to install the qm-wayland
package in the qm_rootfs pipeline. Enabling this
also copies the container image into the final
generated image and a drop-in configuration.

This patch also adds qm_wayland_source to
define the compositor image to something
other than default (which atm is weston).
This offers some flexibility to support
different graphic stacks or compositors for
different purposes.

The way to set this is with simple aib format:

```
experimental:
  internal_defines:
    qm_wayland_enable: true
```

Signed-off-by: Albert Esteve <aesteve@redhat.com>