automotive-image-runner: Change --secureboot to --secureboot-vars
This makes more sense than using `OVMF_VARS.secboot.fd`, as that
contains the microsoft key, which is unlikely to be used in automotive
or embedded use.

If the given file doesn't exist, it is initialized with an empty var
file copy.

There is also a --secureboot-writeable which needs to be passed when
you want to persist to the file, such as when enrolling keys in it.