Add contrib/secureboot
This contains documentation and scripts to help create EFI signing
keys, enroll them and sign EFI binaries with them. It also contains
a pregenerated key and an OVMF_VARS.fd containing the enrolled keys.

Note: This doesn't actually do the work of signing an automotive image
or bootc container. Instead it represents the tasks that a vendor
would do in their own system. That is: deliver a firmware with
enrolled keys, and having the ability to sign EFI binary files with
the secret key.