refactor: Move target board partials to include/ directory
Moves common `ipp.yml` that are not mean to be boards from the `targets/` directory
directly into a new `targets/include/` subdirectory. This improves organization and
clarifies which files are intended for direct use versus those meant to be included.

It also helps with clarity when running the `list-targets` command line option, with
this change only the boards that are intended for direct use are listed.

Signed-off-by: Enric Balletbo i Serra <eballetb@redhat.com>