podman: Add (and use) new PodmanImageMount object
This is a context that makes it easy to use with "with" to ensure the
mount is always paired with an unmount.

This also adds some other method that can be useful, and support
writing to mounts. The methods (potentially) use sudo to allow access
to files in the container that may otherwise not be readable.