fix(main): Use % operator with logging
Use lazy % operator for strings that go into
logging, instead of f-strings.

Logging is optimized to use % formatting [1], so that
formatting operation is deferred until the message
needs to be printed (and only  if the log level allows) [2].

So instead of having to decide when it is worth deferring or not
(i.e., using f-strings or % operators, which can also change in the
future), it is better to just be consistent in how logging is used
across the project.

[1] -
https://docs.python.org/3/howto/logging-cookbook.html#formatting-styles
[2] - https://docs.python.org/3/howto/logging.html#optimization

Signed-off-by: Albert Esteve <aesteve@redhat.com>