Add bootc-reseal operation and bootc-prepare-reseal
If a bootc image was built with composefs signature sealing
(use_composefs_signed), then a changed bootc image (e.g. from
Containerfile layering, or from `bootc-inject-signed`) can be
re-sealed, to allow it to be booted by the initrd in the image.

The simplest way this works is to just run "bootc-reseal" with
no arguments. This will create a new build key and add it to the
initrd, and then sign the image with it and throw away the key.

Unfortunately that doesn't work when you are using signed EFI images,
because adding the build key would modify the initrd and invalidate
any signatures you have on them. In that case you can pre-generate
your own build keys, with something like:

```
$ openssl genpkey -algorithm ed25519 -outform PEM -out private.pem
```

Optionally, you can add `-aes-256-cbc` if you want to password protect
the key.

You can then split out the reseal by calling `bootc-prepare-reseal
--key private.pem`, adding the public key to the initrd, then sign the
initrd, and then complete the seal with `bootc-reseal --key
private.pem`.

Note: This depends on the rpm-ostree changes in:
 https://github.com/coreos/rpm-ostree/pull/5497
 https://github.com/coreos/rpm-ostree/pull/5502

As well as this ostree fix:
 https://github.com/ostreedev/ostree/pull/3527