fix(progress): sudo password prompt swallow
When running a build with sudo, the progress bar
was hiding/swallowing the prompt, which kept
the progress stuck while the user remained
unaware that it was being prompted.

This only happens with rich classes. Fallback
classes were printing the prompt as expected.

To solve this, the only option I found is to
deactivate the auto_refresh from the Progress
instantiation[1], and refresh manually when
anything is printed in stdout.

This makes the refresh less dynamic. The
spinner gets stuck every now and then when the
underlying command does not print to stdout for
a while. ETA is not updated smoothly either.
But overall, this is preferable than having
sudo prompt not visible.

[1]
https://rich.readthedocs.io/en/stable/reference/progress.html#rich.progress.Progress

Signed-off-by: Albert Esteve <aesteve@redhat.com>