Clean up CLI argument parsing code
This makes the code more data driven so it is easier to
maintainer and understand what arguments there are and which are
shared.

I compared the output of:
```
automotive-image-builder --help; for i in list-dist list-targets list-exports compose list-rpms build build-bootc-builder bootc-to-disk-image bootc-extract-for-signing bootc-inject-signed ; do automotive-image-builder  $i --help; done)
```
Before and after this comment, and it is identical.