AIBParameters: add --logfile option that enables with progress
Add --logfile option to specify where the logfile is going
to be created. By default it is
"build_dir/automotive-image-builder-<timestamp>.log".

If progress bar is enabled, this option is always enabled
and will create the log file, either in the default destination
or in the specified through command line.

Otherwise, if progress bar is not enabled, it will only log
to file if explicitely set with --logfile.

Signed-off-by: Albert Esteve <aesteve@redhat.com>