Correct EFI source path in image mode
For some boards, i.e AM69-SK and J784S4XEVM, the first EFI (FAT32) partition
can be used to place the bootloader. This is usually implemented with a
org.osbuild.copy stage in the target board that copies the bootloader or
the firmware from some place to the mount point of the first partition.

By removing the trailing 'EFI', the pipeline now copies all the
'efi' contents, including the 'EFI' directory, ensuring other contents
in the 'efi' directory are properly copied to '/boot/efi'.

Signed-off-by: Enric Balletbo i Serra <eballetb@redhat.com>