aboot: Use the correct aboot.img in image mode
When writing aboot.img to the boot_a partition, in image mode, use
the aboot.img in /usr/lib/modules/$KERNEL_VER/aboot.img instead of
the one in /boot.

This is the "official" path that e.g. osbuild looks aboot.img in when
deploying, and /boot is not even technically part of an ostree commit.
We still use /boot in package mode, though.

This will let us later drop this part in aboot-update:

https://gitlab.com/CentOS/automotive/src/aboot-update/-/blob/0.8.8/aboot-update?ref_type=tags#L141