ridesx4: refactor target definitions for new scmi variants
The scmi variants are still quite volatile and the existing definitions
don't play well with the commonalities between the non-SCMI/SCMI
targets on top of the v2/v3 differences.

Attempt to factorize the configurations around
- ridesx4_common
 - ridesx4
  - ridesx4_r3
 - ridesx4_scmi
  - ridesx4_scmi_r3

Which should cover all use cases and contain the SCMI volatility outside
of the non-SCMI targets. It is also worth noting that the non-SCMI
targets will eventually be phased out due to lack of support.

Signed-off-by: Eric Chanudet <echanude@redhat.com>