utils: Add SudoTemporaryDirectory
This is compatible with tempfile.TemporaryDirectory, but falls
back to cleaning up with sudo rm -rf, which is needed in some
cases when we have osbuild building things with sudo.

Also, there is a detach() call that allows this to be used in
a with clause, but then returned without cleanup in case everything
was successful.