Don't put exports list in the defines
Now that --export is deprecated, lets stop pushing it to the set of
defines. That was always kind of weird, the set of available exports
are part of the generated osbuild manifest and should not affect how
the osbuild manifest is created.

The one place we used this was to see if we're building a bootc image,
but we now handle that by adding a new image_mode of "bootc" instead,
which makes more sense. We the use this to compute use_osbuild and
use_bootc.