Support growing the rootfs using systemd-repart
If the simple manifest has this:

```
image:
  partitions:
    root:
      grow: True
```

Then systemd-repart and systemd-growfs are added to the initrd
and are used to resize the sysroot/rootfs before boot in case
the physical image is larger than the partition table.
So, you can extend this by just truncating a raw image larger, or
by writing it to a larger sdcard.

NOTE: systemd-repart (and thus this feature) only works with GPT
partitioning, and a-i-b will throw an error if this is used on
MBR targets.

Also, this is forbidden by the hardened policy, as live repartitioning
seems like a very risky thing to do in a production environment.