Drop the default value for --cache-max-size
This means that if you don't specify it whatever the current
cache size is, is used. Otherwise we kept reseting it to the
default size, which isn't great.

Additionally, the default osbuild size is "unlimited", and this
is actually a good default for the CI. In fact, we ran into some
issue in the integration_test build in !488, where the
image-size demo failed with:

```
+ automotive-image-builder build --distro=autosd10-sig --cache /var/tmp/tmt/dnf-cache --build-dir /var/tmp/tmt/build --define sign_kernel_modules=false --define reproducible_image=true --target qemu --export qcow2 image-size-2500mb.aib.yml test-image.qcow2
Traceback (most recent call last):
  File "/usr/bin/osbuild", line 33, in <module>
    sys.exit(load_entry_point('osbuild==163', 'console_scripts', 'osbuild')())
             ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/lib/python3.12/site-packages/osbuild/main_cli.py", line 185, in osbuild_cli
    r = manifest.build(
        ^^^^^^^^^^^^^^^
  File "/usr/lib/python3.12/site-packages/osbuild/pipeline.py", line 526, in build
    res = pl.run(store, monitor, libdir, debug_break, stage_timeout)
          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/lib/python3.12/site-packages/osbuild/pipeline.py", line 400, in run
    results = self.build_stages(store,
              ^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/lib/python3.12/site-packages/osbuild/pipeline.py", line 390, in build_stages
    object_store.commit(tree, stage.id)
  File "/usr/lib/python3.12/site-packages/osbuild/objectstore.py", line 437, in commit
    self.cache.store_tree(object_id, obj.path + "/.")
  File "/usr/lib/python3.12/site-packages/osbuild/util/fscache.py", line 1260, in store_tree
    with self.store(name) as rpath_data:
         ^^^^^^^^^^^^^^^^
  File "/usr/lib64/python3.12/contextlib.py", line 144, in __exit__
    next(self.gen)
  File "/usr/lib/python3.12/site-packages/osbuild/util/fscache.py", line 964, in store
    self._remove_lru(info["size"])
  File "/usr/lib/python3.12/site-packages/osbuild/util/fscache.py", line 1131, in _remove_lru
    with self._atomic_open(
         ^^^^^^^^^^^^^^^^^^
  File "/usr/lib64/python3.12/contextlib.py", line 137, in __enter__
    return next(self.gen)
           ^^^^^^^^^^^^^^
  File "/usr/lib/python3.12/site-packages/osbuild/util/fscache.py", line 389, in _atomic_open
    fd = os.open(path, flags, 0o644)
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^
FileNotFoundError: [Errno 2] No such file or directory: '/var/tmp/tmt/build/osbuild_store/objects/5dd9dae3e327c5a711af504e1697ad4e2faec48331511a61226a61db2c4a5003/object.lock'
```

My guess is that this was caused by some other parallel test removing
a checkpointed pipeline. And basically in general, parallel use of a build dir
which has a limited cache size is not really possible due to this issue.

I wish there was a way to set the default cache size without overriding it, becasue
i'm not sure unlimited is great outside of CI runs, but for now I think this is
the best we can do.