Fix creating a root user in a non-simple manifest
If a mpp.yml manifest sets a root password, this is now being
overwritten with the default root password, which is now "not able
to login". This isn't great, especially as the CI qa.mpp.yml does
this.

This changes how we implement the "not able to login" option, which
is still the default to just doing nothing for the root password.
This still means "not able to login" normally, but in the case where
something else set the password we now respect that again.