aboot: Ensure we have a vbmeta file in the image.
This creates an empty vbmeta file in the image and write it to the
vbmeta_a partition as needed. This doesn't anything in the normal
case, because the file is empty, so it will not boot in verified boot
mode. However, having the file being writted to the partition in
bootc-to-disk-image means we can sign the aboot files in the container
image and the resulting image *will* work.

NOTE: For lame technical reasons this isn't empty, as mpp didn't let
me create a zero size file.