contrib: Add example script to sign aboot images
By default this uses the google avb test key from
  https://android.googlesource.com/platform/external/avb/+/refs/tags/android-9.0.0_r37/test/data
which matches the default u-boot avb key.

This can be used to sign the aboot files in a bootc image like this:

automotive-image-builder bootc-extract-for-signing localhost/auto-image to-sign
contrib/avb/sign.sh to-sign/
automotive-image-builder bootc-inject-signed localhost/auto-image localhost/auto-image-signed

And the resulting image from this can be tested with `automotive-image-runner --avb`.

Note: If using sealed image, the abve must be then combined with
prepare-seal and reseal operations.