bootc-reseal: Fix selinux permissions issue
Running aboot-update and bootconfig in a podman container caused
selinux permissions errors when it tried to access things on the
host. The fix is to run these with:
  --security-opt label=type:unconfined_t

However, given that we've forgotten other things in these podman
commandlines, like `--rm`. This insteads creates a shared helper
so we don't have to repeat these arguments all over the place.