arguments: move all args handling to its own module
Create argument.py and move all argument handling to that
new module. In order to avoid circular dependencies, use
a callbacks structure in parse_args with all the functions
that will be used as callbacks filled in main. The rest of
the structure shall basically remain the same.

Adapt the test to work with the new structure.

Signed-off-by: Albert Esteve <aesteve@redhat.com>