arguments: use decorator for command declaration
Add a command() decorator that can be used to fill the
callbacks. By default, use the function __name__ as the
command identifier. Otherwise, use the name received
as decorator's parameter. This will simplify the code,
make it closer to the original and avoid having to fill
and pass the callbacks dictionary.

Signed-off-by: Albert Esteve <aesteve@redhat.com>