Removed supplementary group assignment
The m entry of a sysuser.d file only assigns a supplementary group to
the given user. The primary user is defined in the u entry directly
see here for more details:
https://www.freedesktop.org/software/systemd/man/latest/sysusers.d.html#u
Since none of the users were assigned to a supplementary group prior to
https://gitlab.com/CentOS/automotive/src/automotive-image-builder/-/merge_requests/476
this change needs to be reversed, thus removing all m entries.
In addition, the sysusers.d entries are sorted by ID for better readability.

Signed-off-by: Michael Engel <mengel@redhat.com>