progress: sudo password refresh
Building an image requires sudo commands, but the default timeout
of a sudo command is typically 15 minutes, while a build may in some
cases take longer. So it may happen that when copying the resulting
image to the current folder, the user gets prompted for the password
again, which may cause a problem if the build is left unattended.

Also, the progress bar swallows the output, which required to adjust the
refresh rate, leading to a less responsive bar.

To solve this, whenver the runner runs a sudo command for the first time,
start a thread that periodically runs `sudo -v`. This way we get to:
a) prompt for password before entering the progress bar, and b) refresh
the password timeout limit.

As a result, we can restore the progress bar refresh rate back to normal
and ensure that the password is only requested once, without storing
anything in system memory.

Signed-off-by: Albert Esteve <aesteve@redhat.com>