separate-partitions: Use sparseness when extracting separate partitions.
This uses hole-seeking instead of trailing zeros to detect where a
partition ends. This allows us to detect the case where zeros were
actually written as part of the end of the partition data, which needs
to be written to the real disk.

Also, it changes the extracting code to reproduce the holes in the built
image in the separate partition files.