utils: Add convert_to_simg
This converts regular images to simg files, but it does so without
trying to find zeros, it just leaves out the part of the image that is
really sparse (i.e. has holes in it).

This means anything written to the disk (for example by the mkfs or
the copying of files to the fs), including zeros will be included in
the simg file, and you can safely write such a file to a flash and
just skip writing the sparse data (i.e. no need to write zeros), as
those areas are not going to be referenced by the image.