Fix some permissions with export and --container
osbuild is run with --privileged and is able to override some selinux
labels. This means when we run export commands like img2simg in a
container without the full osbuild permissions we get a selinux error.
So, we introduce another "privileged level" which lets us pass, but
not *all* the privileged.