Make volume handling of args.out more precise
These days, args.out is sometimes not a path, for example when
creating a bootc container. This can cause issues currently because we
always add args.out as a volume.

osbuild anyway will never write to args.out (we move it to the
destination later), so we just remove the addition of args.out as a volume
in _run_osbuild(). Instead we add this only in the precise places we need it.