Assign fixed primary GID for base users in sysusers.d
Relates to: https://gitlab.com/CentOS/automotive/src/automotive-image-builder/-/issues/91

By assigning fixed, primary GIDs to the users in the base-user.conf
systemd-sysusers will not attempt to resolve the primary group  by
searching for a match a match of user and group name. This also
solves the issue of already existing (root) users:
https://gitlab.com/CentOS/automotive/src/automotive-image-builder/-/issues/91
In addition, a 00- prefix to its file name is added. This increases
the "priority" in which the sysuser.d of a-i-b is processed by
systemd-sysusers compared to other sysuser configs such as the one
from setup.rpm:
https://src.fedoraproject.org/rpms/setup

Signed-off-by: Michael Engel <mengel@redhat.com>