build: Allow optionally to build a disk image
This allows a single command to build both a bootc container and
a disk image. This is not always what you want, as you may want
to tweak the bootc container before converting. However, in many
cases, especially during development, it *is* what you want.

Also if the container image name is "-" we don't export the
final container to the image store (or actually, we remove it at the end).
This is going to be useful for testing.