CI: Clean up workflow rules
I ran into an issue with a new MR, this:
```
    - if: '$CI_COMMIT_BRANCH && $CI_OPEN_MERGE_REQUESTS'
      when: never
    - if: '$CI_COMMIT_BRANCH'
```

Intended to run CI when commiting to branches that don't have open
merge requests. However, when doing an initial MR, when pushing the
branch it will not have any open MR, which means all MRs get two
identical CI pipelines started.

The new approach is to instead only run CI for branches `main`
and the release branches like `aib-1.1.x`.

Also, while on this, make it only run for tags that matche
a release format (like `1.1.5`) as well.