aboot: Don't produce an boot-efi.mount file on aboot
For technical reasons we have to create an efi partition to satisfy bootupd
even though it is not used when aboot is active. However, this leads
to problems on when using --separate-partitions on boards that don't
actually have an efi partition, because the boot-efi.mount unit fails
on boot when this partition is missing.

We never ever need this partition anyway, so lets just skip producing
the mount unit file in the workaround case (aboot with no EFI).